First iteration of the new Katib UI (#1427)
* Create a folder for the new-ui

We will create a `new-ui` folder under the `pkg` dir to add the new UI.
This will ensure that we won't break any existing functionality.

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* Initial code for the frontend

This PR introduces the new UI. We hope that this will be the last big PR
in this repo and all of the subsequent ones will be smaller bit-sized
PRs.

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* backend: Expose the entire status of an experiment

We want the table in the main UI page to show more information for each
experiment. This information lives in the status of each Experiment CR,
so we expand the API to also return the entire status for each
Experiment.

In the future we will probably need to just send the entire CR to the
frontend and not parse it at all in the backend.

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* backend: Return the KFP run uid

We want to return the Pipeline UID for a Trial, if such exists.

When combined with Kale, a Trial initiates a KFP run. In this case,
there is an annotation with the KFP run ID, which we can use to navigate
the user to the KFP UI for the specific run.

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* backend: sereve an Angular SPA

To serve an SPA the backend must return the index.html for any non-API
route. The index.html must be sent for any request to the app's page.
Then, once the javascript loads, the app will show to the user the
correct view.

In this commit we also completely remove any caching of the index.html,
for the browser to always request the latest version. This eliminates
the need to hard reload the page to view changes to the frontend code.

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* Dockerfile for the new Katib web app

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* Extend the dockerignore for the new UI

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* Update the README with build commands

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* review: use port 8080 instead of 80 in backend

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* review: use lowercase fields when fetchin exps

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* review: Add seconds to the x-axis of Trial info

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* review: Unify the npm run build commands

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* review: Move TypeMeta values to a common place

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* review: Remove section for max_old_space_size in README

Co-authored-by: Andrey Velichkevich <andrey.velichkevich@gmail.com>

* review: add katib prefix to docs link

Co-authored-by: Andrey Velichkevich <andrey.velichkevich@gmail.com>

* review: Correct link for new UI in README

Co-authored-by: Andrey Velichkevich <andrey.velichkevich@gmail.com>

* review: Remove unused 'format' npm script

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* review: Ensure format checks work with Travis

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* review: Remove unused space

Co-authored-by: Andrey Velichkevich <andrey.velichkevich@gmail.com>

* review: Use create_experiment route

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* review: fix travis govet test

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* review: Rename the Bayesian settings

Co-authored-by: Andrey Velichkevich <andrey.velichkevich@gmail.com>

* review: Rename the ParametersSpec

Co-authored-by: Andrey Velichkevich <andrey.velichkevich@gmail.com>

* review: Remove setting TypeMeta and ObjectMeta

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* review: Update README for build:watch

Co-authored-by: Andrey Velichkevich <andrey.velichkevich@gmail.com>

* review: Fix a typo

Co-authored-by: Andrey Velichkevich <andrey.velichkevich@gmail.com>

* review: Remove unused css

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* review: Use types from k8s.models file

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* Add kfp-run column if UID is present in trials

With Kale a Trial can launch a distinct KF Pipeline. The UID of this
pipeline will be also set as an annotation to the Trial owning the
Pipeline.

In this case the UI should have one extra column for this UID.

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* Properly expose the NAS fields

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* review: Move MetricCollector enums to global enums file

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* review: Remove unused volume enum

Co-authored-by: Andrey Velichkevich <andrey.velichkevich@gmail.com>

* review: Don't send empty settings

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* review: Add parameters for TPE

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

Co-authored-by: Andrey Velichkevich <andrey.velichkevich@gmail.com>