Add missing form fields to new UI (#1463)
* ui(form): Add Trial Parameters

The form will be showing a dynamic list of trial parameters that the
users will need to configure. This list is affected from the yaml
content.

The JS parses the yaml contents to find the trial parameters.

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* ui(form): Refactor the algorithm component

Create a common component for the algorithm settings. We will need this
for Early Stopping, which also has algorithm settings.

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* ui(form): Add support for early stopping

Add a distinct section for Early Stopping, when the Search Algorithm is
for Hyper Parameter tuning.

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* ui(form): Add resume policy to form

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* ui(form): Small fixes to algo's settings

* The algorithm settings got applied in the form only after the user
  selected a different algorithm. The preselected value would not have
  the list of settings assigned once the form loads.
* Use null everywhere for the `random_state` parameter

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* review: Use correct form group

Co-authored-by: Andrey Velichkevich <andrey.velichkevich@gmail.com>

* review: Use LongRunning for resumepolicy default

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* ui(form): Use FormArray variable directly

Using "formGroup.get('trialParameters').controls" in the html results in
an error, during the build process. This is because Angular can't deduce that
the control returned from the get() method is a form array.

We will define a FormArray variable and use it directly instead of
get()ing it from he form group.

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

Co-authored-by: Andrey Velichkevich <andrey.velichkevich@gmail.com>