new-ui: Show the Namespace Selector when app is deployed as standalone (#1483)
* ui(base): Use new commit of the latest common code

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* ui(standalone): Expose namespace dropdown

If the Central Dashboard is not present, then the app will show the
namespace selector and will try to fetch the namespaces. This will allow
the app to work in standalone mode, since the users will be able to
navigate between namespaces even without the central dashboard.

The next steps would be to add authorization checks to the backend to
perform SubjectAccessReviews.

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>